feature: create a dedicated folder for the game logic
