+++ libs/game/main.h
@@ -1,0 +1,10 @@
+
+#ifndef GAME_MAIN_H_INCLUDED
+#define GAME_MAIN_H_INCLUDED
+
+// Root
+#include "./macros.h"
+#include "./methods.h"
+#include "./structs.h"
+
+#endif  // GAME_MAIN_H_INCLUDED
