+++ libs/macros.h
@@ -3,30 +3,6 @@
 #define MACROS_H_INCLUDED
 
 /**
- * @def DASHBOARD_ROWS
- * @brief Defines the number of rows in a grid.
- *
- * This macro is used to define the number of rows in a grid.
- * It is typically used in conjunction with the `DASHBOARD_COLS` macro to define
- * the size of a grid.
- *
- * @warning The value of `DASHBOARD_ROWS` must be a positive integer.
- */
-#define DASHBOARD_ROWS 56
-
-/**
- * @def DASHBOARD_COLS
- * @brief Defines the number of columns in a grid.
- *
- * This macro is used to define the number of columns in a grid.
- * It is typically used in conjunction with the `DASHBOARD_ROWS` macro to define
- * the size of a grid.
- *
- * @warning The value of `DASHBOARD_COLS` must be a positive integer.
- */
-#define DASHBOARD_COLS 110
-
-/**
  * @def MAXIMUM_DELAY
  * @brief Defines the maximum delay in milliseconds.
  *
@@ -47,13 +23,5 @@
  * @warning The value of `MINIMUM_DELAY` must be a positive integer.
  */
 #define MINIMUM_DELAY 0
-
-/**
- * @def NEIGHBORHOOD_RADIUS
- * @brief Defines the radius of the neighborhood.
- *
- * @warning The value of `NEIGHBORHOOD_RADIUS` must be greater or equal to 1.
- */
-#define NEIGHBORHOOD_RADIUS 1
 
 #endif  // MACROS_H_INCLUDED
