+++ todos.md
@@ -1,5 +1,4 @@
 # Falta
 
--   Colocar las macros donde corresponden.
 -   Crear una función para mostrar el juego en _Simple DirectMedia Layer (SDL)_.
--   Mover las funciones y métodos dedicados a actuar sobre una estructura `TGame` a una carpeta propia para asi separar responsabilidades.
+-   Llamar a los archivos `.h` en sus correspondientes `.c`.
